Path stuff, also new icons
Changes to the Paths:
Users are now required to enter the Paths or else they arent able to inject. the programm checks it after rebooting via the menu option or on start of the programm.

Also new icons, and more got translated to german.

The design for GBA, NES and SNES is missing, as is functionality for all 5. Also the TKey feature for NDS (i think dunno), GBA, NES and SNES needs to get added. Also need to translate to german. Oh and as long as i didnt finish CNUSDownloader/YAND there will be still the java dependency cause of JNUSTOOL.